+++ src/frontend/web_application/src/modules/draftMessage/components/RecipientList/presenter.jsx
@@ -272,8 +272,8 @@ class RecipientList extends Component {
   }
 
   render() {
+    // useful ?
     const componentId = uuidV1();
-    const dropdownId = uuidV1();
     const {
       searchResults, className, identity, recipients,
     } = this.props;
@@ -311,7 +311,7 @@ class RecipientList extends Component {
             onBlur={this.handleSearchInputBlur}
           />
           <Dropdown
-            id={dropdownId}
+            className="m-recipient-list__dropdown"
             show={this.state.searchTerms ?
               (searchResults.length > 0 && this.state.searchOpened) : false
             }