+++ src/frontend/web_application/src/modules/encryption/selectors/publicKey.js
@@ -1,0 +1,8 @@
+export const selectKeys = (state, contactId) => {
+  if (state.publicKey && state.publicKey[contactId]) {
+    return state.publicKey[contactId].keys;
+  }
+
+  return null;
+};
+