+++ src/frontend/web_application/src/services/openpgp-manager/api.js
@@ -30,8 +30,8 @@ export function generateKey(name, email, passphrase, options = {}) {
   });
 }
 
-export function getKeyFromASCII(armored) {
-  const { keys } = module.key.readArmored(armored);
+export async function getKeyFromASCII(armored) {
+  const { keys } = await module.key.readArmored(armored);
 
   if (keys && keys.length) {
     return keys[0];