+++ src/frontend/web_application/src/store/selectors/identities.js
@@ -1,0 +1,10 @@
+import { createSelector } from 'reselect';
+
+export const remoteIdentitySelector = state => state.remoteIdentity;
+export const localIdentitySelector = state => state.localIdentity;
+
+export const identitiesSelector = createSelector(
+  localIdentitySelector, remoteIdentitySelector,
+  (localIdentityState, remoteIdentityState) =>
+    [...localIdentityState.localIdentities, ...remoteIdentityState.remoteIdentities]
+);